requirements.txt
discord.py>=2.4.0
google-generativeai
aiofiles
python-dotenv